requirements.txt
Django==1.9.4
django-constance==1.1.2
django-mptt==0.8.4
djangorestframework==3.3.3
Pillow
django-cors-headers==1.3.1
channels
asgi_redis
